+++ .size-snapshot.json
@@ -1,8 +1,8 @@
 {
   "index.js": {
-    "bundled": 4798,
-    "minified": 2055,
-    "gzipped": 909,
+    "bundled": 5027,
+    "minified": 2248,
+    "gzipped": 957,
     "treeshaked": {
       "rollup": {
         "code": 61,
@@ -14,13 +14,32 @@
     }
   },
   "index.cjs.js": {
-    "bundled": 5434,
-    "minified": 2404,
-    "gzipped": 995
+    "bundled": 5779,
+    "minified": 2669,
+    "gzipped": 1054
   },
   "index.iife.js": {
-    "bundled": 5780,
-    "minified": 1888,
-    "gzipped": 870
+    "bundled": 6151,
+    "minified": 1999,
+    "gzipped": 913
+  },
+  "vanilla.js": {
+    "bundled": 3425,
+    "minified": 1526,
+    "gzipped": 697,
+    "treeshaked": {
+      "rollup": {
+        "code": 22,
+        "import_statements": 22
+      },
+      "webpack": {
+        "code": 1045
+      }
+    }
+  },
+  "vanilla.cjs.js": {
+    "bundled": 3923,
+    "minified": 1831,
+    "gzipped": 782
   }
 }
